Move location filter to its own file to make it reusable
